+++ FabrikaLine_Backend/src/main/java/com/example/fabrikaline_backend/Services/ComplaintCategoryServiceImpl.java
@@ -57,7 +57,7 @@ public class ComplaintCategoryServiceImpl implements IComplaintCategoryService, 
     }
 
     @Override
-    public ComplaintCategory getById(int id) {
+    public ComplaintCategory getById(Long id) {
         return null;
     }
 }
